[Feat] #PO-44 - BookProfileModel 파일 분리 및 필드 확장
- Model/BookProfileModel.swift: 신규 파일로 분리, isbn/author/publisher 필드 추가
- BookProfileView.swift: 중복 모델 정의 삭제 (Model/ 파일 참조)
