+++ LivingStory-iOS/LivingStory-iOS/Presentation/Screens/BookProfile/BookProfileView.swift
@@ -1,18 +1,5 @@
 import SwiftUI
 
-// 추후 데미안 ISBN에서 넘겨주는 Book모델로 변경 예정
-struct BookProfileModel: Codable {
-    let bookCoverImageURL: URL?
-    let bookTitle: String
-    let bookDescription: String
-
-    static let mock = BookProfileModel(
-        bookCoverImageURL: URL(string: "https://covers.openlibrary.org/b/isbn/9780156012195-L.jpg"),
-        bookTitle: "완다는 별의 소리를 들어요",
-        bookDescription: "사막에 불시착한 비행사가 별에서 온 왕자를 만나 대화하며, 여우에게 길들여짐과 관계의 소중한 책임감을 배운 뒤에, 사랑하는 장미가 있는 자신의 별로 돌아가는 이야기입니다."
-    )
-}
-
 struct BookProfileView: View {
 
     // MARK: - Properties
